pyproject.toml
[project]
name = "dicto"
version = "2.1.4"
description = "Minimalist desktop app for voice-to-text transcription with global hotkey"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Titovilal", email = "terturionsland@gmail.com"}
]
keywords = ["voice", "transcription", "clipboard", "whisper", "hotkey"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "PySide6>=6.5.0",
    "pynput>=1.7.6",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.0",
    "pyperclip>=1.8.2",
    "httpx>=0.24.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Titovilal/dicto-desktop"
Repository = "https://github.com/Titovilal/dicto-desktop"
Issues = "https://github.com/Titovilal/dicto-desktop/issues"

[project.scripts]
dicto = "src.main:main"

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pytest>=8.0",
    "pytest-qt>=4.3",
    "pytest-mock>=3.12",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "api: tests that hit the real Dicto API (deselect with -m 'not api')",
]
qt_api = "pyside6"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-qt>=4.5.0",
]
